Added file photonIdentification and the program is now in a state that runs. Result is untested though.

+++ source/libcheetah/include/detectorObject.h
@@ -286,7 +286,7 @@ public:
 	double *darkSigmaMap;
 	//float *significanceMap;
 	//float *photonMap;
-	float *cumPhotonMap;
+	long *cumPhotonMap;
 
 	/*
 	 *  Radial stacks for this detector
